requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.1
django-allauth==0.47.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-writable-nested==0.6.3
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==9.3.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.8
whitenoise==6.2.0
